+++ packages/core/scripts/src/docgen/Docgen.ts
@@ -208,7 +208,12 @@ export class Docgen {
     const subComponents = subComponentsNames.reduce((prev: Doc[], curr) => {
       const subComponent = curr.replace(`${this.component.name}.`, "");
       const schema = this.getSchema(`${subComponent}Properties`);
-      const props = this.formatProps(schema);
+      this.component.source = this.getSource(
+        `${this.component.path}/components/${subComponent}/${subComponent}.tsx`,
+      );
+
+      const polymorphicProps = this.getPolymorphicProps();
+      const props = this.formatProps(schema, polymorphicProps);
       prev.push({
         name: curr,
         totalProps: props.length,