+++ packages/implementations/react/src/composite/Breadcrumb/src/components/Separator/index.ts
@@ -1,0 +1,4 @@
+import { Separator } from "./Separator";
+
+export { Separator } from "./Separator";
+export default Separator;