+++ packages/styles/src/packages/composite/breadcrumb/index.ts
@@ -1,0 +1,7 @@
+import { breadcrumbStyles } from "./rarui-breadcrumb.css";
+
+export const breadcrumb = {
+  classnames: breadcrumbStyles,
+};
+
+export type { BreadcrumbVariants } from "./rarui-breadcrumb.types";