+++ packages/styles/src/packages/composite/breadcrumb/rarui-breadcrumb.types.ts
@@ -1,0 +1,6 @@
+import { RecipeVariants } from "@vanilla-extract/recipes";
+import { breadcrumbStyles } from "./rarui-breadcrumb.css";
+
+export type BreadcrumbVariants = RecipeVariants<
+  typeof breadcrumbStyles.breadcrumbItem
+>;